# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api",
        "version": "1.0"
    },
    "paths": {
...
        "/api/values": {
            "get": {
                "tags": [
                    "Attribute Value"
                ],
                "summary": "All Attribute Values",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "attribute_id": {
                                                        "type": "integer",
                                                        "example": "1"
                                                    },
                                                    "product_id": {
                                                        "type": "integer",
                                                        "example": "1"
                                                    },
                                                    "value": {
                                                        "type": "string",
                                                        "example": "Name"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Attribute Value"
                ],
                "summary": "Create Attribute",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "": {
                                                "type": "array",
                                                "items": {
                                                    "properties": {
                                                        "attribute_id": {
                                                            "type": "integer"
                                                        },
                                                        "product_id": {
                                                            "type": "integer"
                                                        },
                                                        "Value": {
                                                            "type": "string"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            },
                            "example": {
                                "attribute_id": 1,
                                "product_id": 1,
                                "value": "Name"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "attribute_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "product_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "value": {
                                                    "type": "string",
                                                    "example": "Name"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/values/{value}": {
            "get": {
                "tags": [
                    "Attribute Value"
                ],
                "summary": "Attribute Value",
                "parameters": [
                    {
                        "name": "value",
                        "in": "path",
                        "description": "Attribute Value ID",
                        "required": true,
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "attribute_id": {
                                                    "type": "integer",
                                                    "example": "3"
                                                },
                                                "product_id": {
                                                    "type": "integer",
                                                    "example": "3"
                                                },
                                                "value": {
                                                    "type": "string",
                                                    "example": "Name"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Attribute Value"
                ],
                "summary": "Attribute Value delete",
                "parameters": [
                    {
                        "name": "value",
                        "in": "path",
                        "description": "Attribute Value ID",
                        "required": true,
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "done"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Attribute Value"
                ],
                "summary": "Attribute Valu Update",
                "parameters": [
                    {
                        "name": "value",
                        "in": "path",
                        "description": "Attribute Value ID",
                        "required": true,
                        "example": 2
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "attribute_id": {
                                                "type": "integer",
                                                "example": "1"
                                            },
                                            "product_id": {
                                                "type": "integer",
                                                "example": "1"
                                            },
                                            "value": {
                                                "type": "string",
                                                "example": "Name"
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "attribute_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "product_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "value": {
                                                    "type": "sting",
                                                    "example": "Name"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}